# Conflicts:
#	master-node-docker/app.py
#	master-node-docker/doc/api_data.json
#	master-node-docker/doc/api_project.js
#	master-node-docker/doc/api_project.json
#	master-node-docker/doc/css/style.css
#	master-node-docker/doc/fonts/glyphicons-halflings-regular.eot
#	master-node-docker/doc/fonts/glyphicons-halflings-regular.svg
#	master-node-docker/doc/fonts/glyphicons-halflings-regular.ttf
#	master-node-docker/doc/fonts/glyphicons-halflings-regular.woff
#	master-node-docker/doc/fonts/glyphicons-halflings-regular.woff2
#	master-node-docker/doc/img/favicon.ico
#	master-node-docker/doc/index.html
#	master-node-docker/doc/locales/ca.js
#	master-node-docker/doc/locales/de.js
#	master-node-docker/doc/locales/es.js
#	master-node-docker/doc/locales/fr.js
#	master-node-docker/doc/locales/it.js
#	master-node-docker/doc/locales/locale.js
#	master-node-docker/doc/locales/nl.js
#	master-node-docker/doc/locales/pl.js
#	master-node-docker/doc/locales/pt_br.js
#	master-node-docker/doc/locales/ro.js
#	master-node-docker/doc/locales/ru.js
#	master-node-docker/doc/locales/tr.js
#	master-node-docker/doc/locales/vi.js
#	master-node-docker/doc/locales/zh.js
#	master-node-docker/doc/locales/zh_cn.js
#	master-node-docker/doc/main.js
#	master-node-docker/doc/utils/handlebars_helper.js
#	master-node-docker/doc/vendor/bootstrap.min.css
#	master-node-docker/doc/vendor/bootstrap.min.js
#	master-node-docker/doc/vendor/diff_match_patch.min.js
#	master-node-docker/doc/vendor/handlebars.min.js
#	master-node-docker/doc/vendor/jquery.min.js
#	master-node-docker/doc/vendor/list.min.js
#	master-node-docker/doc/vendor/lodash.custom.min.js
#	master-node-docker/doc/vendor/path-to-regexp/LICENSE
#	master-node-docker/doc/vendor/path-to-regexp/index.js
#	master-node-docker/doc/vendor/polyfill.js
#	master-node-docker/doc/vendor/prettify.css
#	master-node-docker/doc/vendor/prettify/lang-Splus.js
#	master-node-docker/doc/vendor/prettify/lang-aea.js
#	master-node-docker/doc/vendor/prettify/lang-agc.js
#	master-node-docker/doc/vendor/prettify/lang-apollo.js
#	master-node-docker/doc/vendor/prettify/lang-basic.js
#	master-node-docker/doc/vendor/prettify/lang-cbm.js
#	master-node-docker/doc/vendor/prettify/lang-cl.js
#	master-node-docker/doc/vendor/prettify/lang-clj.js
#	master-node-docker/doc/vendor/prettify/lang-css.js
#	master-node-docker/doc/vendor/prettify/lang-dart.js
#	master-node-docker/doc/vendor/prettify/lang-el.js
#	master-node-docker/doc/vendor/prettify/lang-erl.js
#	master-node-docker/doc/vendor/prettify/lang-erlang.js
#	master-node-docker/doc/vendor/prettify/lang-fs.js
#	master-node-docker/doc/vendor/prettify/lang-go.js
#	master-node-docker/doc/vendor/prettify/lang-hs.js
#	master-node-docker/doc/vendor/prettify/lang-lasso.js
#	master-node-docker/doc/vendor/prettify/lang-lassoscript.js
#	master-node-docker/doc/vendor/prettify/lang-latex.js
#	master-node-docker/doc/vendor/prettify/lang-lgt.js
#	master-node-docker/doc/vendor/prettify/lang-lisp.js
#	master-node-docker/doc/vendor/prettify/lang-ll.js
#	master-node-docker/doc/vendor/prettify/lang-llvm.js
#	master-node-docker/doc/vendor/prettify/lang-logtalk.js
#	master-node-docker/doc/vendor/prettify/lang-ls.js
#	master-node-docker/doc/vendor/prettify/lang-lsp.js
#	master-node-docker/doc/vendor/prettify/lang-lua.js
#	master-node-docker/doc/vendor/prettify/lang-matlab.js
#	master-node-docker/doc/vendor/prettify/lang-ml.js
#	master-node-docker/doc/vendor/prettify/lang-mumps.js
#	master-node-docker/doc/vendor/prettify/lang-n.js
#	master-node-docker/doc/vendor/prettify/lang-nemerle.js
#	master-node-docker/doc/vendor/prettify/lang-pascal.js
#	master-node-docker/doc/vendor/prettify/lang-proto.js
#	master-node-docker/doc/vendor/prettify/lang-r.js
#	master-node-docker/doc/vendor/prettify/lang-rd.js
#	master-node-docker/doc/vendor/prettify/lang-rkt.js
#	master-node-docker/doc/vendor/prettify/lang-rust.js
#	master-node-docker/doc/vendor/prettify/lang-s.js
#	master-node-docker/doc/vendor/prettify/lang-scala.js
#	master-node-docker/doc/vendor/prettify/lang-scm.js
#	master-node-docker/doc/vendor/prettify/lang-sql.js
#	master-node-docker/doc/vendor/prettify/lang-ss.js
#	master-node-docker/doc/vendor/prettify/lang-swift.js
#	master-node-docker/doc/vendor/prettify/lang-tcl.js
#	master-node-docker/doc/vendor/prettify/lang-tex.js
#	master-node-docker/doc/vendor/prettify/lang-vb.js
#	master-node-docker/doc/vendor/prettify/lang-vbs.js
#	master-node-docker/doc/vendor/prettify/lang-vhd.js
#	master-node-docker/doc/vendor/prettify/lang-vhdl.js
#	master-node-docker/doc/vendor/prettify/lang-wiki.js
#	master-node-docker/doc/vendor/prettify/lang-xq.js
#	master-node-docker/doc/vendor/prettify/lang-xquery.js
#	master-node-docker/doc/vendor/prettify/lang-yaml.js
#	master-node-docker/doc/vendor/prettify/lang-yml.js
#	master-node-docker/doc/vendor/prettify/prettify.css
#	master-node-docker/doc/vendor/prettify/prettify.js
#	master-node-docker/doc/vendor/prettify/run_prettify.js
#	master-node-docker/doc/vendor/require.min.js
#	master-node-docker/doc/vendor/semver.min.js
#	master-node-docker/doc/vendor/webfontloader.js
#	master-node-docker/run.sh
#	master-node-docker/sentinel/client/__init__.py
#	master-node-docker/sentinel/client/transactions.py
#	master-node-docker/sentinel/client/validations.py
#	master-node-docker/sentinel/client/vpn.py
#	master-node-docker/sentinel/config.py
#	master-node-docker/sentinel/eth/contracts.py
#	master-node-docker/sentinel/helpers/__init__.py
#	master-node-docker/sentinel/helpers/eth.py
#	master-node-docker/sentinel/node/payments.py
#	smart-contracts/services/VPNService.sol
#	vpn-node-docker/Dockerfile.dev
#	vpn-node-docker/app.py
#	vpn-node-docker/sentinel/config.py
#	vpn-node-docker/sentinel/node/controllers.py
#	wallet/electron.js
#	wallet/package.json
#	wallet/public/index.html
#	wallet/src/Actions/AccountActions.js
#	wallet/src/Components/Create.js
#	wallet/src/Components/Dashboard.js
#	wallet/src/Components/EtherTransaction.js
#	wallet/src/Components/Header.js
#	wallet/src/Components/History.js
#	wallet/src/Components/Home.js
#	wallet/src/Components/ReceiveComponent.js
#	wallet/src/Components/SendComponent.js
#	wallet/src/Components/SentTransaction.js


Former-commit-id: 9f16b09f8c2edbb42a5b3aa31b07459724ddbeb1

+++ wallet/src/Actions/AccountActions.js
@@ -2,7 +2,8 @@ const fs = window.require('fs');
 const electron = window.require('electron');
 const remote = electron.remote;
 const { exec } = window.require('child_process');
-const B_URL = 'http://35.198.220.210:8000';
+const B_URL = 'http://35.198.246.114:8000';
+const MASTER_URL = 'http://35.198.204.28:8000'
 const ETH_BALANCE_URL = `https://api.etherscan.io/api?apikey=Y5BJ5VA3XZ59F63XQCQDDUWU2C29144MMM
 &module=account&action=balance&tag=latest&address=`;
 const SENT_BALANCE_URL = `https://api.etherscan.io/api?apikey=Y5BJ5VA3XZ59F63XQCQDDUWU2C29144MMM
@@ -12,6 +13,8 @@ const ETH_TRANSC_URL = `https://api.etherscan.io/api?apikey=Y5BJ5VA3XZ59F63XQCQD
 const SENT_TRANSC_URL1 = `https://api.etherscan.io/api?apikey=Y5BJ5VA3XZ59F63XQCQDDUWU2C29144MMM
 &module=logs&action=getLogs&fromBlock=0&toBlock=latest&address=0xa44E5137293E855B1b7bC7E2C6f8cD796fFCB037
 &topic0=0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef&topic0_1_opr=and&topic1=`;
+const TRANSC_STATUS = `https://api.etherscan.io/api?&apikey=Y5BJ5VA3XZ59F63XQCQDDUWU2C29144MMM&
+module=transaction&action=gettxreceiptstatus&txhash=`;
 const SENT_TRANSC_URL2 = `&topic1_2_opr=or&topic2=`;
 const SENT_DIR = getUserHome() + '/.sentinel';
 const KEYSTORE_FILE = SENT_DIR + '/keystore';
@@ -99,6 +102,52 @@ export function getAccount(cb) {
   });
 }
 
+export function transferAmountMaster(data, cb) {
+  getKeystore(function (err, keystore) {
+    if (err) cb(err, null);
+    else {
+      data['keystore'] = keystore;
+      fetch(MASTER_URL + '/client/transaction', {
+        method: 'POST',
+        headers: {
+          'Accept': 'application/json',
+          'Content-type': 'application/json',
+          'Access-Control-Allow-Origin': '*'
+        },
+        body: JSON.stringify(data)
+      }).then(function (response) {
+        response.json().then(function (response) {
+          console.log("Response...", response)
+          if (response.success === true) {
+            var tx_hash = response['tx_hash'];
+            cb(null, tx_hash);
+          } else {
+            cb({ message: response.message || 'Error occurred while initiating transfer amount.' }, null);
+          }
+        })
+      });
+    }
+  });
+}
+
+export function getTransactionStatus(tx_addr, cb) {
+  fetch(TRANSC_STATUS + tx_addr, {
+    method: 'GET',
+    headers: {
+      'Accept': 'application/json',
+      'Content-type': 'application/json',
+      'Access-Control-Allow-Origin': '*'
+    }
+  }).then(function (response) {
+    response.json().then(function (response) {
+      console.log("Response...", response)
+      if (response.status === "1") {
+        var status = response['result']["status"];
+        cb(null, status);
+      } else cb({ message: 'Error occurred while getting balance.' }, null);
+    })
+  });
+}
 
 export function transferAmount(data, cb) {
   getKeystore(function (err, keystore) {
@@ -114,14 +163,19 @@ export function transferAmount(data, cb) {
         },
         body: JSON.stringify(data)
       }).then(function (response) {
-        response.json().then(function (response) {
-          if (response.success === true) {
-            var tx_hash = response['tx_hash'];
-            cb(null, tx_hash);
-          } else {
-            cb({ message: 'Error occurred while initiating transfer amount.' }, null);
-          }
-        })
+        if (response.status === 200) {
+          response.json().then(function (response) {
+            if (response.success === true) {
+              var tx_hash = response['tx_hash'];
+              cb(null, tx_hash);
+            } else {
+              cb({ message: response.message || 'Error occurred while initiating transfer amount.' }, null);
+            }
+          })
+        }
+        else {
+          cb({ message: 'Error occurred while initiating transfer amount.' }, null);
+        }
       });
     }
   });
@@ -137,7 +191,7 @@ export function getEthBalance(data, cb) {
     }
   }).then(function (response) {
     response.json().then(function (response) {
-      if (response.status == '1') {
+      if (response.status === '1') {
         var balance = response['result'] / (10 ** 18);
         cb(null, balance);
       } else cb({ message: 'Error occurred while getting balance.' }, null);
@@ -180,7 +234,7 @@ export function getEthTransactionHistory(account_addr, page, cb) {
 }
 
 export function getSentTransactionHistory(account_addr, cb) {
-  fetch(SENT_TRANSC_URL1 + account_addr + SENT_TRANSC_URL2+account_addr, {
+  fetch(SENT_TRANSC_URL1 + account_addr + SENT_TRANSC_URL2 + account_addr, {
     method: 'GET',
     headers: {
       'Accept': 'application/json',
@@ -192,6 +246,26 @@ export function getSentTransactionHistory(account_addr, cb) {
         var history = response['result'];
         cb(null, history);
       } else cb({ message: 'Error occurred while getting transaction history.' }, null);
+    });
+  });
+}
+
+export function getVpnHistory(account_addr, cb) {
+  fetch(B_URL + '/client/vpn/usage', {
+    method: 'POST',
+    headers: {
+      'Accept': 'application/json',
+      'Content-type': 'application/json'
+    },
+    body: JSON.stringify({
+      account_addr: account_addr,
+    })
+  }).then(function (response) {
+    response.json().then(function (response) {
+      if (response.success === true) {
+        var history = response['usage'];
+        cb(null, history);
+      } else cb({ message: response.message || 'Error occurred while getting vpn history.' }, null);
     });
   });
 }
@@ -214,39 +288,58 @@ export const getVPNList = (cb) => {
   });
 }
 
+
 export function connectVPN(account_addr, vpn_addr, cb) {
   var command = 'sudo openvpn ' + OVPN_FILE;
-
-  getOVPNAndSave(account_addr, vpn_addr, function (err) {
-    if (err) cb(err);
-    else {
-      if (OVPNDelTimer) clearInterval(OVPNDelTimer);
-
-      exec(command, function (err, stdout, stderr) {
-        OVPNDelTimer = setTimeout(function () {
-          fs.unlinkSync(OVPN_FILE);
-        }, 5 * 1000);
-      });
-
-      setTimeout(function () {
-        getVPNPIDs(function (err, pids) {
+  fetch(B_URL + '/client/vpn', {
+    method: 'POST',
+    headers: {
+      'Accept': 'application/json',
+      'Content-type': 'application/json',
+    },
+    body: JSON.stringify({
+      account_addr: account_addr,
+      vpn_addr: vpn_addr
+    })
+  }).then(function (response) {
+    response.json().then(function (res) {
+      if (res.success === true) {
+        getOVPNAndSave(account_addr, res['ip'], res['port'], vpn_addr, res['token'], function (err) {
           if (err) cb(err);
           else {
-            CONNECTED = true;
-            cb(null);
+            if (OVPNDelTimer) clearInterval(OVPNDelTimer);
+
+            exec(command, function (err, stdout, stderr) {
+              OVPNDelTimer = setTimeout(function () {
+                fs.unlinkSync(OVPN_FILE);
+              }, 5 * 1000);
+            });
+
+            setTimeout(function () {
+              getVPNPIDs(function (err, pids) {
+                if (err) cb(err);
+                else {
+                  CONNECTED = true;
+                  cb(null);
+                }
+              });
+
+            }, 1000);
           }
         });
-
-      }, 1000);
-    }
-  });
+      }
+      else {
+        cb({ message: res.message || 'Error occurred while connecting vpn.' }, null);
+      }
+    })
+  })
 }
 
-function getOVPNAndSave(account_addr, vpn_addr, cb) {
+function getOVPNAndSave(account_addr, vpn_ip, vpn_port, vpn_addr, nonce, cb) {
   if (fs.existsSync(OVPN_FILE)) {
     cb(null);
   } else {
-    fetch(B_URL + '/client/vpn', {
+    fetch('http:' + vpn_ip + ':' + vpn_port + '/client/generateOVPN', {
       method: 'POST',
       headers: {
         'Accept': 'application/json',
@@ -254,7 +347,8 @@ function getOVPNAndSave(account_addr, vpn_addr, cb) {
       },
       body: JSON.stringify({
         account_addr: account_addr,
-        vpn_addr: vpn_addr
+        vpn_addr: vpn_addr,
+        token: nonce
       })
     }).then(function (response) {
       response.json().then(function (response) {
@@ -286,6 +380,15 @@ export function getVPNPIDs(cb) {
       cb(true, null);
     }
   });
+}
+
+export const isOnline = function () {
+  if (window.navigator.onLine) {
+    return true
+  }
+  else {
+    return false
+  }
 }
 
 export function disconnectVPN(cb) {
